Renamed BSONObjectID to ObjectID. Removed BSONObjectID product.
