+++ Sources/BSONCompose/Values/ObjectID+ValueProtocol.swift
@@ -5,7 +5,7 @@
 //  Created by Christopher Richez on April 16 2022
 //
 
-import BSONObjectID
+import ObjectID
 
 extension ObjectID: ValueProtocol {
     public var bsonType: UInt8 { 7 }
